backend_python/requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
torch
ffmpeg-python
python-dotenv